requirements.txt
django_extensions
jupyter
pydot
graphviz
pandas
matplotlib
